+++ src/dataoperation.rs
@@ -27,8 +27,7 @@ use log::debug;
 
 struct CounterNonceSequence(u32);
 
-//extern crate lsb_text_png_steganography_mod;
-//use lsb_text_png_steganography_mod::{ hide, reveal_mod };
+use crate::lsb_text_png_steganography_mod::{ hide_mod, reveal_mod };
 
 impl NonceSequence for CounterNonceSequence {
     // called once for each seal operation
@@ -50,6 +49,7 @@ pub enum DataOperation {
     WEBPAGE,
     ROT13, //only after base64 because input is String
     REVERSE,
+    STEGANO(Stegano),
 }
 
 pub trait UnApplyDataOperation {
@@ -78,6 +78,9 @@ pub trait UnApplyDataOperation {
         };
         Ok(caps["loader"].as_bytes().to_vec())
     }
+
+
+
 }
 impl UnApplyDataOperation for DataOperation {
     fn un_apply_one_operation(&self, data: Vec<u8>) -> Result<Vec<u8>, anyhow::Error> {
@@ -86,6 +89,7 @@ impl UnApplyDataOperation for DataOperation {
             DataOperation::ROT13 => self.rot13_decode(data),
             DataOperation::WEBPAGE => self.webpage_harvesting(data),
             DataOperation::AEAD(aead_material) => aead_material.decrypt_mat(data),
+            DataOperation::STEGANO(stegano) => stegano.stegano_decode_lsb(data),
             _ => todo!(),
         }
     }
@@ -106,6 +110,7 @@ pub trait ApplyDataOperation {
         debug!("{}", encrypt_string!("dataoperation: WEBPAGE create"));
         Ok(format!("!!!{}!!!", std::str::from_utf8(&data)?).into_bytes())
     }
+
 }
 impl ApplyDataOperation for DataOperation {
     fn apply_one_operation(&mut self, data: Vec<u8>) -> Result<Vec<u8>, anyhow::Error> {
@@ -114,8 +119,32 @@ impl ApplyDataOperation for DataOperation {
             DataOperation::ROT13 => self.rot13_encode(data),
             DataOperation::WEBPAGE => self.webpage_create(data),
             DataOperation::AEAD(aead_material) => aead_material.encrypt_mat(data),
+            DataOperation::STEGANO(stegano) => stegano.stegano_encode_lsb(data),
             _ => todo!(),
         }
+    }
+}
+
+#[derive(Serialize, Deserialize, Debug, PartialEq, Clone)]
+pub struct Stegano {
+    pub input_image: String,
+}
+impl Stegano {
+    fn stegano_decode_lsb(&self, data: Vec<u8>) -> Result<Vec<u8>, anyhow::Error> {
+        debug!("{}", encrypt_string!("dataoperation: STEGANO decode"));
+        Ok(reveal_mod(data)?)
+    }
+
+    fn stegano_encode_lsb(&self, data: Vec<u8>) -> Result<Vec<u8>, anyhow::Error> {
+        debug!("{}", encrypt_string!("dataoperation: STEGANO encode"));
+        //Ok(hide_mod(data,&self.input_image)?)
+        let img = hide_mod(data, self.input_image.as_str());
+        //img.save('output_carrier_path').unwrap();
+        // TODO il faut pas laisser ça !!!
+        debug!("IMAGE SAVE to /tmp/output_rust.png");
+        img.save("/tmp/output_rust.png").unwrap();
+        Ok(img.to_vec())
+        //todo!()
     }
 }
